% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{ate_reg}
\alias{ate_reg}
\title{function to compute ate using regression adjustment}
\usage{
ate_reg(est, w, y, df, fml, ret_est = T, ...)
}
\arguments{
\item{est}{string estimator name, must be \%in\% c("ols", "logit",  "lasso", "ridge", "rforest", "xgboost")}

\item{df}{data.frame}

\item{fml}{formula for outcome model (cannot include treatment)}

\item{ret_est}{boolean for whether to return ATE estimate, returns yhat vectors when false}

\item{name}{of outcome}
}
\description{
function to compute ate using regression adjustment
}
