% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{ehatter}
\alias{ehatter}
\title{fit pscore using specified algorithm}
\usage{
ehatter(estimator, fmla, df, w, ...)
}
\arguments{
\item{estimator}{string with name of pscore estimator}

\item{fmla}{formula for conditional mean estimator}

\item{df}{dataframe}

\item{y}{name of outcome variable}
}
\value{
pscore model object
}
\description{
fit pscore using specified algorithm
}
