% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{fit_me}
\alias{fit_me}
\title{Ingredients for AIPW estimator}
\usage{
fit_me(meanfn, pscorefn, mean_fml, psc_fml, y, w, df, ...)
}
\arguments{
\item{meanfn}{name of outcome model}

\item{pscorefn}{name of pscore model}

\item{mean_fml}{formula for outcome model}

\item{psc_fml}{formula for pscore model}

\item{y}{outcome name}

\item{w}{treatment name}

\item{df}{dataframe}
}
\description{
Ingredients for AIPW estimator
}
