% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{fit_me_xf}
\alias{fit_me_xf}
\title{Ingredients for AIPW estimator (cross-fit)}
\usage{
fit_me_xf(meanfn, pscorefn, mean_fml, psc_fml, y, w, df, nfold = 2L, ...)
}
\arguments{
\item{meanfn}{name of outcome model}

\item{pscorefn}{name of pscore model}

\item{mean_fml}{formula for outcome model}

\item{psc_fml}{formula for pscore model}

\item{y}{outcome name}

\item{w}{treatment name}

\item{df}{dataframe}

\item{nfold}{number of folds (2 by default)}
}
\description{
Ingredients for AIPW estimator (cross-fit)
}
