% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{mhatter}
\alias{mhatter}
\title{fit CEF using specified algorithm}
\usage{
mhatter(estimator, fmla, df, y, ...)
}
\arguments{
\item{estimator}{string with name of conditional mean estimator}

\item{fmla}{formula for conditional mean estimator}

\item{df}{dataframe}

\item{y}{name of outcome variable (required for packages that don't use formulas to fit)}
}
\value{
outcome model object
}
\description{
fit CEF using specified algorithm
}
